장병문 : Y정렬로 몬스터, 눈물, 플레이어 위치정렬 - 플레이어 머리, 몸 위치는 정해서 하도록 해야됨

+++ Portfolio/GameEngineContents/ShopItem.cpp
@@ -28,7 +28,7 @@ void ShopItem::SetType(ShopType _Type)
 		break;
 	case ShopType::Passive:
 	{
-		int RandomInt = GameEngineRandom::MainRandom->RandomInt(1, PassiveItem::ItemNameVector_.size() - 1);
+		int RandomInt = GameEngineRandom::MainRandom->RandomInt(1, static_cast<int>(PassiveItem::ItemNameVector_.size() - 1));
 		std::string TmpName = PassiveItem::ItemNameVector_[RandomInt];
 		PassiveType_ = static_cast<PassiveType>(RandomInt);
 		GameEngineRenderer* TmpRenderer = CreateRenderer(TmpName);
